OMAP4: dpll: Control async bridge to ABE and L3/MA to EMIF
On OMAP4460 when the MPU Frequency is above 748Mhz, the programmable
divider for the Async bridge to ABE must be set to MPU-Freq/8. For
lower frequency, it should be MPU-Freq/4.

Similarly for MPU Frequency above 920Mhz, the programmable divider
for the async bridge to L3 and Memory Adapter interfaces of EMIF
must be MPU-Freq/4. For lower frequency, they should be MPU-Freq/2.

Do this as part of the DPLL control api.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>