MFD: palmas PMIC device support
Palmas is a PMIC from Texas Instruments and this is the MFD part of the
driver for this chip. The PMIC has SMPS and LDO regulators, a general
purpose ADC, GPIO, USB OTG mode detection, watchdog and RTC features.

Contains WIP code to start to support sharing the rtc-twl as its
the same IP block

Signed-off-by: Graeme Gregory <gg@slimlogic.co.uk>

Conflicts:

	drivers/mfd/Kconfig
	drivers/mfd/Makefile