ARM: OMAP5: clockdata: Add alias for DMIC
Create an alias for the DMIC FCLK as 'dmic_fck'.
OMAP4 colck data uses dmic_fck for the functional clock name.
In order to get the DMIC clock parenting working we need to
have consistent names for the functional clock.

Create an alias for the PAD CLK as 'pad_clks_ck'.
OMAP4 colck data uses pad_clks_ck for the clock name.
In order to get the DMIC clock parenting working we need to
have consistent names for the pad clock.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Sebastien Guiriec <s-guiriec@ti.com>