ASoC: dpcm: Add DAI only PCM ops
Some DSP component drivers need to be able to configure specific DAIs
for certain use cases. Provide a public DAI PCM operations API in order
to support this.

This is used atm on the OMAP4 ABE MODEM port which can use the McBSP2
DAI to directly perform IO on an ABE FE PCM (mutually exclusive).

Signed-off-by: Liam Girdwood <lrg@ti.com>