Input: twl6040-vibra: Update for device tree support
The twl6040 DT support implementation has been changed from the originally
planned.
None of the child devices going to have compatible_of property which means
that the child devices of twl6040 will be created as traditional MFD devices.
The mfd core driver will decide (based on the DT blob) to create a device
for the twl6040-vibra or not. If the DT blob has 'vibra' section the device
will be created without pdata.
In this case the vibra driver will reach up to the parent node to get the
needed properties.
With DT booted kernel we no longer be able to link the regulators to the vibra
driver, they can be only linked to the MFD device (probed via DT). From the
vibra driver we ned to use pdev->dev.parent to get the regulators.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>