ASoC: twl6040: HACK: Accessory type detection
With the functionality test mode we should be able to detect the
type of accessory pluged in.
The theory goes:
we receive the plug interrupt for insertation of accessory
we start a functionality test.
we check the HSMICBIAS connection test result
if it is passed we have headset, if it fails we have headphone.

However:
We receive the insertation interrupt too early when the plug is not
fully in. [1]
If we follow the TRM recommendation we detect everything as headphone.
The headset most of the time been detected as headphone. It seams waiting
for about a minute between unplug/plug sequence helps on Blaze.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>