ASoC: omap-abe-twl6040: McASP support update
With this commit we should be fine when booting with DT and in Legacy mode
with the updated omap-mcasp driver.

If we boot with DT and the ti,mcasp phandle is not present we skip registering
the dai link for McASP.
In Legacy boot we always enable the McASP (lazy to modify the pdata structure).

Can be squashed if you can find the relevant commit.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>