ARM: OMAP4+: HDMI: Refine the DMA port resource for audio
Instead of passing the complete address space to the platform device for audio,
just pass the address offset of the DMA port for audio samples. Thus, we prevent
that two drivers try to ioremap the same resources. This is to be safe, as the
ASoC HDMI CPU-DAI driver will not need to ioremap such resource.

Signed-off-by: Ricardo Neri <rneri@dextratech.com>