ASoC: omap-mcpdm: Drop using devm_request_threaded_irq()
We need to free the irq at omap_mcpdm_remove() which is called when the
machine driver has been removed (the card has been removed).
If we fail to do that, next time when the machine driver is loaded the
dai's probe will fail since the irq has been already requested.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>