ASoC: davinci-evm: HACK Beagle X15 support HACK
Handle the Beagle X15 with separate compatible:
Since MCLK can not be provided to the codec we are not compatible with any
of the other configurations.
However we can use BCLK instead of MCLK for codec's internals. This will
produce 'some sort of' audio but it is far from perfect.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>