drm/omap: Fix race in HPD detection
It was possible to hit a race condition when the HPD detection was enabled,
but the DRM stack was still not fully configured.

By separating the HPD callback registration and the HPD enable call we can
move the HPD enable to a location when the DRM stack is up already, thus
avoiding the race.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>