drm/i2c: tda998x: Reset the I2S_FORMAT (Page0, Reg 0xfc) to it's default
The reset value of the register is 0, the soft reset does not reset this
register and if other kernel changed this the audio is going to be
distorted.

It was observed when - accidentally - booted the kernel from eMMC on BBB
which is 3.8.13-bone79 and it sets this register to 0x0a. After reboot and
tda998x_reset() it remains 0x0a.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>