ASoC: ti: davinci-mcasp: Support for serializer reconfiguration
Currently the serializer direction can be set only once and it is static.
If the same McASP is connected to different codecs with different sets of
serializers, there is a need to change the serializer pins depending on
which codec is being used currently.

With the new function machine drivers can feed new serial_dir array if
needed.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>