arm64: dts: ti: k3-j721e-common-proc-board: Move refclk2 pinctrl to k3_clks
The audio_refclk2 used as output from j721e to the pcm3168a codec as SCKI.

While it is fine to have the pin attached to the codec, but it is actually
a clock pin.
The two codec on the IVI board uses refclk0 as their SCKI and for them the
pinctrl must be in the k3_clks node - one pinctrl can not be shared between
the two codec.

So move the REFCLK2 pinctrl to k3_clks as first thing.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>